spring mvc 2

+++ library-web-app-spring/src/main/java/com/company/librarywebappspring/service/impl/BookServiceImpl.java
@@ -4,7 +4,6 @@ import com.company.librarywebappspring.models.Book;
 import com.company.librarywebappspring.repository.BookRepository;
 import com.company.librarywebappspring.service.inter.BookService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
@@ -18,14 +17,23 @@ public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
 
     @Override
-    public List<Book> findAll() {
-        return bookRepository.findAll();
+    public List<Book> findAll(String search) {
+        return bookRepository.findAllByParam(search);
     }
 
     @Override
     public boolean save(Book book) {
         bookRepository.save(book);
         return true;
+    }
+
+    @Override
+    public boolean update(Book book) {
+        if(bookRepository.existsById(book.getId())){
+            bookRepository.save(book);
+            return true;
+        }
+        return false;
     }
 
     @Override
@@ -36,7 +44,10 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public boolean remove(Integer id) {
-        bookRepository.deleteById(id);
-        return true;
+        if(bookRepository.existsById(id)){
+            bookRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
